requirements.txt
cryptography==37.0.4
pycryptodome==3.17
requests==2.28.1
